fix(pointcloud): near-term batch — correctness + code-org
Four items from issue #611's near-term list:

1. computeBBox empty + non-finite guards (e57 + ifcx-points). Empty
   or NaN-poisoned chunks no longer produce ±Infinity bounds.

2. Magic-byte-first format detection. detectPointCloudFormat now
   probes the buffer before falling back to extension, so a LAS
   file named *.ply still loads correctly.

3. E57 packet-bounds + per-stream guards. Validate that the
   DataPacket header, length table, and each bytestream fit inside
   payloadEnd before reading.

4. Code-org splits:
   - e57.ts (631 lines) → e57-page.ts + e57-xml.ts + e57-decode.ts
     + e57.ts orchestrator (all <400 lines)
   - point-cloud-renderer.ts (420 lines) → extract uniform writer
     to point-cloud-uniforms.ts (renderer drops to 364 lines)

Verified: 62 pointcloud unit tests, full repo typecheck (24/24).
